daos completed, code reformatted, implementation of NavigationDrawer sta...

+++ Vkazam/src/main/java/com/github/spitsinstafichuk/vkazam/daos/FingerprintDao.java
@@ -23,10 +23,12 @@ public class FingerprintDao {
     @SuppressWarnings("unused")
     public Fingerprint get(long id) {
         SQLiteDatabase db = databaseHelper.getReadableDatabase();
-        assert db != null;
+        if (db == null) {
+            return null;
+        }
         Cursor cursor = db.query(FingerprintTable.FINGERPRINTS, null, FingerprintTable._ID + "=?",
-                new String[] {
-                    Long.toString(id)
+                new String[]{
+                        Long.toString(id)
                 }, null, null, null);
         Fingerprint fingerprint = null;
         if (cursor.moveToFirst()) {
@@ -35,7 +37,8 @@ public class FingerprintDao {
                     cursor.getString(cursor.getColumnIndex(FingerprintTable.FINGERPRINT)),
                     new Date(
                             cursor.getLong(cursor.getColumnIndex(FingerprintTable.DATE))),
-                    new Date(cursor.getLong(cursor.getColumnIndex(FingerprintTable.DELETION_DATE))));
+                    new Date(
+                            cursor.getLong(cursor.getColumnIndex(FingerprintTable.DELETION_DATE))));
         }
         db.close();
         return fingerprint;
@@ -44,7 +47,9 @@ public class FingerprintDao {
     @SuppressWarnings("unused")
     public Fingerprint getLastFingerprint() {
         SQLiteDatabase db = databaseHelper.getReadableDatabase();
-        assert db != null;
+        if (db == null) {
+            return null;
+        }
         Cursor cursor = db.query(FingerprintTable.FINGERPRINTS, null, null,
                 null, null, FingerprintTable.DATE + " DESC", "1");
         Fingerprint fingerprint = null;
@@ -54,7 +59,8 @@ public class FingerprintDao {
                     cursor.getString(cursor.getColumnIndex(FingerprintTable.FINGERPRINT)),
                     new Date(
                             cursor.getLong(cursor.getColumnIndex(FingerprintTable.DATE))),
-                    new Date(cursor.getLong(cursor.getColumnIndex(FingerprintTable.DELETION_DATE))));
+                    new Date(
+                            cursor.getLong(cursor.getColumnIndex(FingerprintTable.DELETION_DATE))));
         }
         db.close();
         return fingerprint;
@@ -64,7 +70,13 @@ public class FingerprintDao {
     public long save(Fingerprint fingerprint) {
         long result;
         SQLiteDatabase db = databaseHelper.getWritableDatabase();
-        assert db != null;
+        if (db == null) {
+            if (fingerprint.getId() == -1) {
+                return -1;
+            } else {
+                return 0;
+            }
+        }
         ContentValues values = new ContentValues();
         values.put(FingerprintTable.FINGERPRINT, fingerprint.getFingerprint());
         values.put(FingerprintTable.DATE, fingerprint.getDate().getTime());
@@ -74,8 +86,8 @@ public class FingerprintDao {
         } else {
             result = db.update(FingerprintTable.FINGERPRINTS, values, FingerprintTable._ID
                     + "=?",
-                    new String[] {
-                        Long.toString(fingerprint.getId())
+                    new String[]{
+                            Long.toString(fingerprint.getId())
                     });
         }
         db.close();
@@ -85,10 +97,12 @@ public class FingerprintDao {
     @SuppressWarnings("unused")
     public int delete(long id) {
         SQLiteDatabase db = databaseHelper.getWritableDatabase();
-        assert db != null;
+        if (db == null) {
+            return 0;
+        }
         int result = db.delete(FingerprintTable.FINGERPRINTS, FingerprintTable._ID + "=?",
-                new String[] {
-                    Long.toString(id)
+                new String[]{
+                        Long.toString(id)
                 });
         db.close();
         return result;
